Removed some usages of OpenGL outside the engine. Fixed the position of the 3D sound listener at the beginning of the game. Removed the spawning sounds for units present at the beginning of the game. Changed the implementation of conditional OpenTK references. Works but causes warnings. Added a little delay before the scrollers scroll when using the mouse.

+++ Sources/GameLogic/Pathfinding/PathFinder.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System;
 using System.Collections;
 using System.Diagnostics;
 using OpenTK.Math;
@@ -107,7 +106,7 @@ namespace Orion.GameLogic.Pathfinding
             visitedNodeCount = 0;
             nodeNearestToDestinationIndex = -1;
 
-            if (openNodeStateValue == byte.MaxValue - 2)
+            if ((int)openNodeStateValue == (int)byte.MaxValue - 2)
             {
                 openNodeStateValue = 1;
                 closedNodeStateValue = 2;
